Add category assignments to product filters

+++ admin-api/src/modules/productFilters/services/productFilters.service.ts
@@ -5,6 +5,7 @@ import { ProductFilterDto } from '@/src/libs/models/dtos/productsFilter/ProductF
 import { ProductFilterEntity } from '@/src/libs/models/entities/productFilter/ProductFilter.entity';
 import { ProductFilterTranslationsEntity } from '@/src/libs/models/entities/productFilter/ProductFilterTranslation.entity';
 import { ProductFilterTypeEnum } from '@/src/utils/enums/ProductFilterEnums';
+import { CategoryEntity } from '@/src/libs/models/entities/category/Category.entity';
 
 @Injectable()
 export class ProductFiltersService {
@@ -16,32 +17,49 @@ export class ProductFiltersService {
   ) {}
 
   async createProductFilter(data: ProductFilterDto) {
-    const nameExists = await this.productFilterRepository.find({
-      where: {
-        translations: {
-          name: In(data.translations.map((t) => t.name)),
+    await this.productFilterRepository.manager.transaction(async (manager) => {
+      const filterRepository = manager.getRepository(ProductFilterEntity);
+      const categoryRepository = manager.getRepository(CategoryEntity);
+      const nameExists = await filterRepository.find({
+        where: {
+          translations: {
+            name: In(data.translations.map((translation) => translation.name)),
+          },
         },
-      },
-    });
-    if (nameExists.length > 0) {
-      throw new BadRequestException('Product filter name must be unique');
-    }
-    const newFilter = this.productFilterRepository.create(data);
+      });
+      if (nameExists.length > 0) {
+        throw new BadRequestException('Product filter name must be unique');
+      }
 
-    await this.productFilterRepository.save(newFilter);
+      const categories = await categoryRepository.find({
+        where: { id: In(data.categoryIds) },
+      });
+      if (categories.length !== data.categoryIds.length) {
+        throw new BadRequestException(
+          'One or more selected categories do not exist',
+        );
+      }
+
+      const newFilter = filterRepository.create({
+        type: data.type,
+        translations: data.translations,
+        categories,
+      });
+      await filterRepository.save(newFilter);
+    });
 
     return HttpStatus.CREATED;
   }
 
   async getProductFilters() {
     return await this.productFilterRepository.find({
-      relations: ['translations'],
+      relations: ['translations', 'categories', 'categories.translations'],
     });
   }
 
   async getProductFilter(id: string) {
     return await this.productFilterRepository.findOne({
-      relations: ['translations'],
+      relations: ['translations', 'categories', 'categories.translations'],
       where: { id },
     });
   }
@@ -52,17 +70,24 @@ export class ProductFiltersService {
       const translationRepo = manager.getRepository(
         ProductFilterTranslationsEntity,
       );
+      const categoryRepo = manager.getRepository(CategoryEntity);
 
       const productFilter = await productFilterRepo.findOneOrFail({
         where: { id },
         lock: { mode: 'pessimistic_write' },
       });
 
-      Object.keys(data).forEach((key) => {
-        if (key !== 'translations') {
-          productFilter[key] = data[key];
-        }
+      const categories = await categoryRepo.find({
+        where: { id: In(data.categoryIds) },
       });
+      if (categories.length !== data.categoryIds.length) {
+        throw new BadRequestException(
+          'One or more selected categories do not exist',
+        );
+      }
+
+      productFilter.type = data.type;
+      productFilter.categories = categories;
 
       await translationRepo.delete({ productFilterId: id });
 
@@ -97,9 +122,16 @@ export class ProductFiltersService {
     }));
   }
 
-  async getProductFiltersWithOptions() {
+  async getProductFiltersWithOptions(categoryId?: string) {
     return await this.productFilterRepository.find({
-      relations: ['translations', 'options', 'options.translations'],
+      where: categoryId ? { categories: { id: categoryId } } : {},
+      relations: [
+        'translations',
+        'options',
+        'options.translations',
+        'categories',
+        'categories.translations',
+      ],
     });
   }
 }
